:sparkles: Create useCodeViewerState hooks.
🎉 - :tada: - Begin a project.
✨ - :sparkles: - Introduce new features.
🔥 - :fire: - Remove code or files.
🐛 - :bug: - Fix a bug.
✏️  - :pencil2: - Fix typos.
⚡️ - :zap: - Improve performance.
📝 - :memo: - Add or update documentation.
🚀 - :rocket: - Deploy stuff.
💄 - :lipstick: - Add or update the UI and style files.
✅ - :white_check_mark: - Add or update tests.
🔒 - :lock: - Fix security issues.
🔖 - :bookmark: - Release / Version tags.
🚧 - :construction: - Work in progress.
💚 - :green_heart: - Fix CI Build.
♻️  - :recycle: - Refactor code.
🔧 - :wrench: - Add or update configuration files.
💩 - :poop: - Write bad code that needs to be improved.
